Pa 07 create landing page web app

+++ frontend/src/app/shared/service/auth.service.ts
@@ -1,10 +1,13 @@
-import { Injectable } from "@angular/core";
-import { BehaviorSubject, Observable, throwError } from "rxjs";
-import { User } from "../model/user";
-import { HttpClient, HttpHeaders } from "@angular/common/http";
-import { catchError, map } from "rxjs/operators";
-import { Router } from "@angular/router";
-import { environment } from "../../../environments/environment";
+import {HostListener, Injectable} from "@angular/core";
+import {BehaviorSubject, Observable, throwError} from "rxjs";
+import {User} from "../model/user";
+import {HttpClient, HttpHeaders} from "@angular/common/http";
+import {catchError, map} from "rxjs/operators";
+import {Router} from "@angular/router";
+import {environment} from "../../../environments/environment";
+import {SignUpRequest} from "../model/signUpRequest";
+import {LoginRequest} from "../model/loginRequest";
+import {ForgotPasswordRequest} from "../model/forgotPasswordRequest";
 
 @Injectable({
   providedIn: 'root'
@@ -12,28 +15,47 @@ import { environment } from "../../../environments/environment";
 export class AuthService {
 
   public user: Observable<User> | undefined;
+  isLoggedIn = false;
   private userSubject!: BehaviorSubject<User>;
 
-  constructor(private router: Router, private http:HttpClient) {
+  constructor(private router: Router, private _http: HttpClient) {
     const user = localStorage.getItem('user');
     if (user !== null) {
-    this.userSubject = new BehaviorSubject<User>( 
-      (this.user = JSON.parse(localStorage.getItem("user") ?? ''))
-    );
-    this.user = this.userSubject.asObservable();
+      this.userSubject = new BehaviorSubject<User>(
+        (this.user = JSON.parse(localStorage.getItem("user") ?? ''))
+      );
+      this.user = this.userSubject.asObservable();
+    }
   }
-   }
-   
 
-   public get userValue(): User {
+  public get userValue(): User {
     return this.userSubject.value;
   }
 
-  login(user: User): Observable<User> {
-    return this.http.post<User>(`${environment.apiUrl}auth/login/`, user).pipe(
+  registerUser(signUpRequest: SignUpRequest) {
+    const url = window.location.href;
+    const headers = new HttpHeaders()
+      .set('X-URL', url);
+    console.log(url);
+    return this._http.post(`${environment.apiUrl}auth/signup`, signUpRequest, {headers});
+  }
+
+  verifyRegistration(verificationCode: string) {
+    return this._http.get(`${environment.apiUrl}auth/signup/verify?code=${verificationCode}`);
+  }
+
+  login(loginRequest: LoginRequest, keepMeSignIn: boolean): Observable<User> {
+    return this._http.post<User>(`${environment.apiUrl}auth/signin`, loginRequest).pipe(
       map((user) => {
-        localStorage.setItem("user", JSON.stringify(user));
-        this.userSubject.next(user);
+        if (keepMeSignIn) {
+          localStorage.setItem("user", JSON.stringify(user));
+          if(this.userSubject !== undefined){
+            this.userSubject.next(user);
+          } else {
+            this.userSubject = new BehaviorSubject<User>(user);
+          }
+        }
+        this.isLoggedIn = true
         return user;
       })
     );
@@ -41,28 +63,29 @@ export class AuthService {
 
   logout(): any {
     localStorage.removeItem("user");
-    this.router.navigate(["/login"]);
+    this.isLoggedIn = false;
+    this.router.navigate(['/']);
   }
-
-  register(user: any) {
-    return this.http.post(`${environment.apiUrl}/auth/register/`, user);
+  forgotPassword(forgotPasswordRequest: ForgotPasswordRequest) {
+    return this._http.post(`${environment.apiUrl}auth/forgotPassword`, forgotPasswordRequest);
   }
 
   getAll(): Observable<User[]> {
-    return this.http.get<User[]>(`${environment.apiUrl}user`);
+
+    return this._http.get<User[]>(`${environment.apiUrl}user`);
   }
 
   getById(id: string): any {
-    return this.http.get<User>(`${environment.apiUrl}/user/${id}`);
+    return this._http.get<User>(`${environment.apiUrl}/user/${id}`);
   }
 
   update(id: string, params: any): any {
-    return this.http.put(`${environment.apiUrl}/user/${id}`, params).pipe(
+    return this._http.put(`${environment.apiUrl}/user/${id}`, params).pipe(
       map((x) => {
         // update stored user if the logged in user updated their own record
         if (id == this.userValue.id) {
           // update local storage
-          const user = { ...this.userValue, ...params };
+          const user = {...this.userValue, ...params};
           localStorage.setItem("user", JSON.stringify(user));
 
           // publish updated user to subscribers
@@ -74,7 +97,7 @@ export class AuthService {
   }
 
   delete(id: string): any {
-    return this.http.delete(`${environment.apiUrl}/user/${id}`).pipe(
+    return this._http.delete(`${environment.apiUrl}/user/${id}`).pipe(
       map((x) => {
         // auto logout if the logged in user deleted their own record
         if (id == this.userValue.id) {
@@ -97,7 +120,7 @@ export class AuthService {
       "Content-Type": "application/json",
     });
 
-    return this.http
+    return this._http
       .request<T>(verb, url, {
         body,
         headers: myHeaders,
@@ -107,9 +130,5 @@ export class AuthService {
           throwError(`Network Error: ${error.statusText} (${error.status})`)
         )
       );
-  }
-
-  RegisterUser(inputdata:any){
-    return this.http.post(`${environment.apiUrl}`,inputdata)
   }
 }
